fix: Renamed jwt.guard file and updated routes into depend files
