+++ src/cart/controller/cart.controller.ts
@@ -25,7 +25,7 @@ import {
 import { UserDto } from 'src/users/dtos/response/user.dto';
 import { Roles } from '../../auth/decorators/roles.decorator';
 import { GetUser } from '../../auth/decorators/user.decorator';
-import { JwtAuthGuard } from '../../auth/guards/jwt-guard';
+import { JwtAuthGuard } from '../../auth/guards/jwt.guard';
 import { Role } from '../../auth/roles.enum';
 import { CartItemsDto } from '../dtos/response/cart-item.dto';
 import { CartDto } from '../dtos/response/cart.dto';
