+++ src/desserts/desserts.module.ts
@@ -4,7 +4,7 @@ import { FilesService } from 'src/files/service/files.service';
 import { LikesModule } from 'src/likes/likes.module';
 import { LikesService } from 'src/likes/service/likes.service';
 import { AuthModule } from '../auth/auth.module';
-import { JwtAuthGuard } from '../auth/guards/jwt-guard';
+import { JwtAuthGuard } from '../auth/guards/jwt.guard';
 import { RolesGuard } from '../auth/guards/roles.guard';
 import { PrismaService } from '../prisma.service';
 import { DessertsController } from './controller/desserts.controller';
